+++ build.gradle
@@ -155,8 +155,8 @@ sortClassFields {
     add 'main', 'org.spongepowered.api.effect.potion.PotionEffectTypes'
     add 'main', 'org.spongepowered.api.effect.sound.SoundTypes'
     add 'main', 'org.spongepowered.api.entity.EntityTypes'
-    add 'main', 'org.spongepowered.api.entity.ai.task.AITaskTypes'
-    add 'main', 'org.spongepowered.api.entity.ai.GoalTypes'
+    add 'main', 'org.spongepowered.api.ai.task.AITaskTypes'
+    add 'main', 'org.spongepowered.api.ai.GoalTypes'
     add 'main', 'org.spongepowered.api.entity.living.complex.dragon.phase.EnderDragonPhaseTypes'
     add 'main', 'org.spongepowered.api.entity.living.player.gamemode.GameModes'
     add 'main', 'org.spongepowered.api.event.cause.EventContextKeys'
