+++ src/main/java/org/spongepowered/api/GameRegistry.java
@@ -26,8 +26,7 @@ package org.spongepowered.api;
 
 import org.spongepowered.api.data.type.Profession;
 import org.spongepowered.api.data.value.ValueFactory;
-import org.spongepowered.api.entity.ai.task.AITaskType;
-import org.spongepowered.api.entity.ai.task.AbstractAITask;
+import org.spongepowered.api.ai.task.AITaskType;
 import org.spongepowered.api.entity.living.Agent;
 import org.spongepowered.api.item.merchant.TradeOfferGenerator;
 import org.spongepowered.api.item.merchant.VillagerRegistry;
@@ -259,19 +258,6 @@ public interface GameRegistry {
      *      <tt>Optional.empty()</tt> if not found
      */
     Optional<DisplaySlot> getDisplaySlotForColor(TextColor color);
-
-    /**
-     * Registers a new {@link AbstractAITask} with an {@link Agent} as the
-     * owner. The complete id will be in the format of
-     * <code>{@link PluginContainer#getId()}:id</code>.
-     *
-     * @param plugin The plugin who owns it
-     * @param id The id that represents the task type
-     * @param name The name for the task
-     * @param aiClass The class of the task
-     * @return The type
-     */
-    AITaskType registerAITaskType(PluginContainer plugin, String id, String name, Class<? extends AbstractAITask<? extends Agent>> aiClass);
 
     /**
      * Gets the {@link ValueFactory} for creating values.
