+++ src/main/java/org/spongepowered/api/ai/package-info.java
@@ -23,4 +23,4 @@
  * THE SOFTWARE.
  */
 @org.spongepowered.api.util.annotation.NonnullByDefault
-package org.spongepowered.api.entity.ai.task.builtin.creature.target;
+package org.spongepowered.api.ai;
