+++ src/main/java/org/spongepowered/api/ai/task/AITask.java
@@ -22,16 +22,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.entity.ai.task;
+package org.spongepowered.api.ai.task;
 
-import org.spongepowered.api.entity.ai.Goal;
+import org.spongepowered.api.ai.Goal;
 import org.spongepowered.api.entity.living.Agent;
 
 import java.util.Optional;
 
 /**
- * Represents a task performed by {@link Agent}s. If you desire to create your own, see
- * {@link AbstractAITask}.
+ * Represents a task performed by {@link Agent}s.
  *
  * @param <O> The type of agent
  */
@@ -57,32 +56,8 @@ public interface AITask<O extends Agent> {
      * @return The owner or {@link Optional#empty()} if not present
      */
     default Optional<O> getOwner() {
-        return getGoal().map(Goal::getOwner);
+        return this.getGoal().map(Goal::getOwner);
     }
-
-    /**
-     * Tests if the provided {@link AITask} is allowed to be ran concurrently
-     * with this task.
-     *
-     * <p>This plays a role in determining if tasks should be updated or not.
-     * If this method
-     * is being called on this task, that means:</p>
-     *
-     * <ol>
-     *   <li>This task has higher priority than the provided task for our
-     *   {@link AITask#getGoal()}.</li>
-     *   <li>Returning "false" will remove the provided task from the list of
-     *   updated tasks, if not
-     *   there already.</li>
-     *   <li>Returning "true" will add the provided task to the list of updated
-     *   tasks, if not there
-     *   already.</li>
-     * </ol>
-     *
-     * @param other The other task
-     * @return True if it can be, false if not
-     */
-    boolean canRunConcurrentWith(AITask<O> other);
 
     /**
      * Returns if this task can be interrupted. This determines if this task
@@ -97,11 +72,11 @@ public interface AITask<O extends Agent> {
      *
      * <p>Due note that the meaning of "true" changes based on the state of the
      * {@link Goal}.  To put it simply, this value can mean "Should I be added
-     * as an updating task or "Should I continue updating?". Regardless of the
-     * question, the next step that happens is transferring from this method to
-     * {@link AITask#canRunConcurrentWith(AITask)} in the next AI tick loop.</p>
+     * as an updating task or "Should I continue updating?".</p>
      *
      * @return True if can be interrupted, false if not
      */
     boolean canBeInterrupted();
+
+    AICategoryFlag getCategoryFlags();
 }
