+++ src/main/java/org/spongepowered/api/ai/task/AITaskTypes.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.entity.ai.task;
+package org.spongepowered.api.ai.task;
 
 import org.spongepowered.api.entity.Entity;
 import org.spongepowered.api.entity.living.Agent;
@@ -65,5 +65,6 @@ public final class AITaskTypes {
     // SORTFIELDS:OFF
 
     private AITaskTypes() {
+        throw new AssertionError("You should not be attempting to instantiate this class.");
     }
 }
