+++ src/main/java/org/spongepowered/api/ai/task/package-info.java
@@ -23,4 +23,4 @@
  * THE SOFTWARE.
  */
 @org.spongepowered.api.util.annotation.NonnullByDefault
-package org.spongepowered.api.entity.ai.task.builtin.creature;
+package org.spongepowered.api.ai.task;
